+++ hive/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Updates suggested placement of `GenerateAdapters` annotation
 - Fixes custom objects in `BoxCollection`. Custom objects must now be json serializable.
 - Adds much more information to the unknown typeId error message
+- Supports `freezed` default values
 
 ## 2.10.1